Consolidate VibesMP library files
