--- builds/MicroPython/apps_unfrozen/vibesmp_lib/help/es.txt
@@ -1,9 +1,0 @@
-VibesMP es una aplicación de audio MicroPython dedicada, diseñada para microcontroladores RP2350. Proporciona una reproducción de MP3 de alta fidelidad aprovechando las capacidades de decodificación acelerada por hardware, manteniendo una baja huella de memoria y una alta capacidad de respuesta durante las operaciones simultáneas de archivos en la tarjeta SD y las tareas de renderizado de la interfaz de usuario en plataformas de hardware integradas.
-
-La aplicación implementa el análisis de etiquetas ID3v2 para la extracción de metadatos, el escalado y la visualización en tiempo real de portadas de álbumes en formato JPEG y BMP, la lógica de transición de reproducción sin interrupciones, la gestión dinámica de listas de reproducción incluyendo creación y modificación, y un sistema de navegación de tres columnas para la exploración eficiente de la biblioteca, la selección de pistas y el cambio de lista de reproducción a través de diversos directorios de almacenamiento.
-
-Los parámetros de configuración disponibles incluyen la alternancia de reproducción automática para el avance secuencial de pistas, la selección de modos de reproducción aleatoria y en bucle, la localización de idiomas para inglés, alemán, francés y español, ajustes visuales preestablecidos mediante paletas RGB565, intervalos de búsqueda ajustables de uno a quince segundos, gestión del tiempo de espera de enfoque y almacenamiento persistente de todas las preferencias.
-
-Los controles de navegación utilizan el mapeo estándar del D-Pad: arriba y abajo para el desplazamiento vertical de la lista, izquierda y derecha para el cambio de columna y el ajuste de volumen, el botón central para la confirmación de la selección y la alternancia de reproducción y pausa, el botón de retroceso para la regresión de vista o la terminación de la aplicación, y un mapeo alfanumérico especializado para el nombramiento.
-
-Realizado por Slasher006 con la ayuda de Gemini, 2026. Este proyecto muestra técnicas avanzadas de implementación de MicroPython, estrategias eficientes de gestión de memoria para sistemas integrados con recursos limitados y el desarrollo colaborativo de software entre ingenieros humanos e inteligencia artificial para ofrecer una experiencia multimedia superior en la plataforma del microcontrolador RP2350.
