--- builds/MicroPython/apps_unfrozen/vibesmp_lib/help/first_start_de.txt
@@ -1,1 +1,0 @@
-Willkommen bei VibesMP! Um Ihre musikalische Reise zu beginnen, führen Sie bitte einen ersten Scan Ihrer SD-Karte durch, um alle verfügbaren MP3-Dateien zu indizieren. Wenn Sie später neue Titel hinzufügen, verwenden Sie einfach die Option "Bibliothek scannen" im Hauptmenü, um Ihre Datenbank zu aktualisieren und sicherzustellen, dass alle neuen Tracks korrekt erkannt und in der Anwendung angezeigt werden.
