--- builds/MicroPython/apps_unfrozen/vibesmp_lib/help/first_start_es.txt
@@ -1,1 +1,0 @@
-¡Bienvenido a VibesMP! Para comenzar su viaje musical, realice un escaneo inicial de su tarjeta SD para indexar todos los archivos MP3 disponibles. Si agrega pistas nuevas más tarde, simplemente use la opción "Escanear biblioteca" en el menú principal para actualizar su base de datos y asegurarse de que toda la música nueva se detecte y se muestre correctamente dentro de la aplicación.
