--- builds/MicroPython/apps_unfrozen/vibesmp_lib/help/fr.txt
@@ -1,9 +1,0 @@
-VibesMP est une application audio MicroPython dédiée, conçue pour les microcontrôleurs RP2350. Elle offre une lecture MP3 haute fidélité en exploitant les capacités de décodage accélérées par le matériel, tout en maintenant une faible empreinte mémoire et une grande réactivité lors des opérations simultanées sur les fichiers de la carte SD et des tâches de rendu de l'interface utilisateur sur les plateformes matérielles embarquées.
-
-L'application implémente l'analyse des balises ID3v2 pour l'extraction des métadonnées, la mise à l'échelle et l'affichage en temps réel des pochettes d'album au format JPEG et BMP, la logique de transition pour la lecture sans interruption, la gestion dynamique des listes de lecture comprenant la création et la modification, et un système de navigation à trois colonnes pour l'exploration efficace de la bibliothèque, la sélection des pistes et le changement de liste de lecture à travers divers répertoires de stockage.
-
-Les paramètres de configuration disponibles incluent l'activation de la lecture automatique pour l'avancement séquentiel des pistes, la sélection du mode aléatoire ou en boucle, la localisation linguistique pour l'anglais, l'allemand, le français et l'espagnol, les thèmes visuels prédéfinis utilisant des palettes RGB565, des intervalles de recherche réglables de une à quinze secondes, la gestion du délai de mise au point et le stockage persistant de toutes les préférences.
-
-Les commandes de navigation utilisent le mappage standard du D-Pad : haut et bas pour le défilement vertical des listes, gauche et droite pour le changement de colonne et le réglage du volume, le bouton central pour la confirmation de la sélection et le basculement lecture et pause, le bouton retour pour le retour à la vue précédente ou l'arrêt de l'application, et un mappage alphanumérique spécialisé pour le nommage.
-
-Réalisé par Slasher006 avec l'aide de Gemini, 2026. Ce projet met en évidence des techniques avancées d'implémentation MicroPython, des stratégies de gestion de mémoire efficaces pour les systèmes embarqués à ressources limitées, et un développement logiciel collaboratif entre des ingénieurs humains et l'intelligence artificielle pour offrir une expérience multimédia supérieure sur la plateforme du microcontrôleur RP2350.
