--- builds/MicroPython/apps_unfrozen/vibesmp_lib/lang/de.json
@@ -1,52 +1,0 @@
-{
-  "menu_player": "Player",
-  "menu_playlist": "Wiedergabeliste",
-  "menu_play_file": "Datei abspielen",
-  "menu_library": "Bibliothek",
-  "menu_editor": "Editor",
-  "menu_settings": "Einstellungen",
-  "menu_help": "Hilfe",
-  "menu_playlist_manager": "Listen-Manager",
-  "now_playing": "Wird abgespielt",
-  "no_track": "Kein Titel",
-  "playing": "Play",
-  "paused": "Pause",
-  "stopped": "Stop",
-  "loop_none": "Wiederholung: Aus",
-  "loop_one": "Wiederholung: Eins",
-  "loop_all": "Wiederholung: Alle",
-  "editor_color": "Farbe aendern",
-  "editor_bg": "Hintergrund",
-  "editor_save": "Speichern",
-  "set_autoplay": "Auto-Play",
-  "set_shuffle": "Zufall",
-  "set_auto_expand": "Bibl. auto-ausklappen",
-  "set_lang": "Sprache",
-  "set_volume": "Lautstaerke",
-  "set_seek": "Sprungweite",
-  "set_focus_timeout": "Fokus Timeout",
-  "refresh_library": "Bibl. aktualisieren",
-  "back": "Zurueck",
-  "on": "AN",
-  "off": "AUS",
-  "menu_playlist_editor": "Listen-Editor",
-  "menu_playlists": "Wiedergabelisten",
-  "library": "Bibliothek",
-  "playlist": "Titel",
-  "playlists": "Listen",
-  "new_playlist": "+ Neue Liste",
-  "playlist_new": "Neue Liste",
-  "playlist_del": "Liste loeschen",
-  "playlist_selector": "Liste waehlen",
-  "playlist_editor": "Listen-Editor",
-  "hint_np_controls": "TAB:Listen P:Pause ESC:Stop [/:Seek <>:Titel L:Loop S:Mix V:.,",
-  "hint_np_lib": "TAB:Player UD:Scroll LR:Bereich OK:Add/Auf",
-  "hint_np_trk": "TAB:Player UD:Scroll LR:Bereich OK:Play DEL:Entf",
-  "hint_np_pls": "TAB:Player UD:Scroll LR:Bereich OK:Laden N:Neu DEL:Loesch",
-  "hint_playlist_sel": "UD:Scroll OK:Laden DEL:Loeschen BACK:Menue",
-  "hint_playlist_ed": "LR:Bereich UD:Wahl OK:Add/Entf BACK:Menue",
-  "hint_continue": "ENT:Weiter",
-  "hint_confirm": "LR:Waehlen ENT:Bestaetigen",
-  "confirm": "Bestaetigen",
-  "delete": "Loeschen"
-}
