--- builds/MicroPython/apps_unfrozen/vibesmp_lib/lang/en.json
@@ -1,92 +1,0 @@
-{
-  "app_name": "VibesMP",
-  "menu_player": "Player",
-  "menu_playlist": "Playlist",
-  "menu_play_file": "Play File",
-  "menu_library": "Library",
-  "menu_editor": "Editor",
-  "menu_settings": "Settings",
-  "menu_help": "Help",
-  "first_run_title": "VibesMP",
-  "first_run_msg": "Scan SD card for music now?",
-  "scanning_title": "Scanning SD Card...",
-  "scan_complete_title": "Scan Complete",
-  "scan_complete_msg": "Found {} MP3 files.",
-  "menu_playlist_manager": "Playlist Manager",
-  "now_playing": "Now Playing",
-  "no_track": "No Track",
-  "playing": "Playing",
-  "paused": "Paused",
-  "stopped": "Stopped",
-  "loop_none": "Loop: No",
-  "loop_one": "Loop: One",
-  "loop_all": "Loop: All",
-  "editor_color": "Change Color",
-  "editor_bg": "Background",
-  "editor_save": "Save",
-  "set_autoplay": "Auto-Play Next",
-  "set_shuffle": "Shuffle",
-  "set_auto_expand": "Auto-Expand Lib",
-  "set_lang": "Language",
-  "set_volume": "Volume",
-  "set_seek": "Seek Step",
-  "set_focus_timeout": "Focus Timeout",
-  "refresh_library": "Refresh Library",
-  "lib_all_songs": "All Songs",
-  "lib_artists": "Artists",
-  "lib_albums": "Albums",
-  "lib_folders": "Folders",
-  "lib_genres": "Genres",
-  "lib_recently_added": "Recently Added",
-  "lib_favorites": "Favorites",
-  "lib_search": "Search",
-  "lib_scan_options": "Scan Options",
-  "lib_sort": "Sort",
-  "lib_filters": "Filters",
-  "lib_stats": "Library Stats",
-  "lib_cleanup": "Cleanup",
-  "lib_scan": "Scan Library",
-  "lib_actions": "Actions",
-  "lib_action_play_now": "Play Now",
-  "lib_action_play_next": "Play Next",
-  "lib_action_add_current": "Add to Current Playlist",
-  "lib_action_create_playlist": "Create Playlist",
-  "lib_action_remove_library": "Remove from Library",
-  "lib_action_add_favorite": "Add Favorite",
-  "lib_action_remove_favorite": "Remove Favorite",
-  "lib_action_show_info": "Show Info",
-  "lib_info": "Track Info",
-  "lib_tracks": "Tracks",
-  "lib_scan_total": "Total",
-  "lib_scan_added": "Added",
-  "lib_scan_removed": "Removed",
-  "lib_scan_unchanged": "Unchanged",
-  "lib_scan_found": "Found",
-  "lib_scan_failed": "Failed",
-  "lib_removed": "Removed",
-  "lib_favorites_cleared": "Favorites cleared",
-  "hint_library": "UD:Scroll LR:Folder OK:Open/Action BACK:Back",
-  "back": "Back",
-  "on": "ON",
-  "off": "OFF",
-  "menu_playlist_editor": "Playlist Editor",
-  "menu_playlists": "Playlists",
-  "library": "Library",
-  "playlist": "Tracks",
-  "playlists": "Playlists",
-  "new_playlist": "+ New Playlist",
-  "playlist_new": "New Playlist",
-  "playlist_del": "Delete Playlist",
-  "playlist_selector": "Select Playlist",
-  "playlist_editor": "Playlist Editor",
-  "hint_np_controls": "TAB:Lists P:Pause ESC:Stop [/:Seek <>:Trk L:Loop S:Shuf V:.,",
-  "hint_np_lib": "TAB:Player UD:Scroll LR:Panels OK:Add/Expand",
-  "hint_np_trk": "TAB:Player UD:Scroll LR:Panels OK:Play DEL:Rem",
-  "hint_np_pls": "TAB:Player UD:Scroll LR:Panels OK:Load N:New DEL:Del",
-  "hint_playlist_sel": "UD:Scroll OK:Load DEL:Delete BACK:Menu",
-  "hint_playlist_ed": "LR:Panel UD:Sel OK:Add/Rem BACK:Menu",
-  "hint_continue": "ENT:Continue",
-  "hint_confirm": "LR:Choose ENT:Confirm",
-  "confirm": "Confirm",
-  "delete": "Delete"
-}
