--- builds/MicroPython/apps_unfrozen/vibesmp_lib/lang/es.json
@@ -1,52 +1,0 @@
-{
-  "menu_player": "Reproductor",
-  "menu_playlist": "Lista",
-  "menu_play_file": "Reproducir archivo",
-  "menu_library": "Biblioteca",
-  "menu_editor": "Editor",
-  "menu_settings": "Ajustes",
-  "menu_help": "Ayuda",
-  "menu_playlist_manager": "Gestor de listas",
-  "now_playing": "Reproduciendo",
-  "no_track": "Sin pista",
-  "playing": "Reproduciendo",
-  "paused": "Pausa",
-  "stopped": "Detenido",
-  "loop_none": "Bucle: No",
-  "loop_one": "Bucle: Uno",
-  "loop_all": "Bucle: Todo",
-  "editor_color": "Cambiar color",
-  "editor_bg": "Fondo",
-  "editor_save": "Guardar",
-  "set_autoplay": "Auto-reproducir",
-  "set_shuffle": "Aleatorio",
-  "set_auto_expand": "Auto-expandir bibl.",
-  "set_lang": "Idioma",
-  "set_volume": "Volumen",
-  "set_seek": "Salto",
-  "set_focus_timeout": "Tiempo Foco",
-  "refresh_library": "Actualizar bibl.",
-  "back": "Atrás",
-  "on": "ON",
-  "off": "OFF",
-  "menu_playlist_editor": "Editor de listas",
-  "menu_playlists": "Listas",
-  "library": "Biblioteca",
-  "playlist": "Pistas",
-  "playlists": "Listas",
-  "new_playlist": "+ Nueva lista",
-  "playlist_new": "Nueva lista",
-  "playlist_del": "Borrar lista",
-  "playlist_selector": "Seleccionar lista",
-  "playlist_editor": "Editor de listas",
-  "hint_np_controls": "TAB:Listas P:Pause ESC:Stop [/:Seek <>:Pista L:Loop S:Shuf V:.,",
-  "hint_np_lib": "TAB:Repro UD:Scroll LR:Panel OK:Add/Abrir",
-  "hint_np_trk": "TAB:Repro UD:Scroll LR:Panel OK:Repro DEL:Borrar",
-  "hint_np_pls": "TAB:Repro UD:Scroll LR:Panel OK:Cargar N:Nueva DEL:Borrar",
-  "hint_playlist_sel": "UD:Scroll OK:Cargar DEL:Borrar BACK:Menu",
-  "hint_playlist_ed": "LR:Panel UD:Sel OK:Add/Rem BACK:Menu",
-  "hint_continue": "ENT:Continuar",
-  "hint_confirm": "LR:Elegir ENT:Confirmar",
-  "confirm": "Confirmar",
-  "delete": "Borrar"
-}
