--- builds/MicroPython/apps_unfrozen/vibesmp_lib/lang/fr.json
@@ -1,52 +1,0 @@
-{
-  "menu_player": "Lecteur",
-  "menu_playlist": "Liste",
-  "menu_play_file": "Lire fichier",
-  "menu_library": "Bibliothèque",
-  "menu_editor": "Éditeur",
-  "menu_settings": "Réglages",
-  "menu_help": "Aide",
-  "menu_playlist_manager": "Gestionnaire",
-  "now_playing": "En lecture",
-  "no_track": "Aucun titre",
-  "playing": "Lecture",
-  "paused": "Pause",
-  "stopped": "Arrêté",
-  "loop_none": "Boucle: Non",
-  "loop_one": "Boucle: Un",
-  "loop_all": "Boucle: Tout",
-  "editor_color": "Changer couleur",
-  "editor_bg": "Fond",
-  "editor_save": "Enregistrer",
-  "set_autoplay": "Lecture auto",
-  "set_shuffle": "Aléatoire",
-  "set_auto_expand": "Auto-développer bibl.",
-  "set_lang": "Langue",
-  "set_volume": "Volume std",
-  "set_seek": "Saut",
-  "set_focus_timeout": "Delai Focus",
-  "refresh_library": "Actualiser bibl.",
-  "back": "Retour",
-  "on": "ON",
-  "off": "OFF",
-  "menu_playlist_editor": "Éditeur de listes",
-  "menu_playlists": "Listes",
-  "library": "Bibliothèque",
-  "playlist": "Titres",
-  "playlists": "Listes",
-  "new_playlist": "+ Nouvelle liste",
-  "playlist_new": "Nouvelle liste",
-  "playlist_del": "Supprimer liste",
-  "playlist_selector": "Choisir liste",
-  "playlist_editor": "Éditeur de listes",
-  "hint_np_controls": "TAB:Listes P:Pause ESC:Stop [/:Seek <>:Titre L:Boucle S:Rand V:.,",
-  "hint_np_lib": "TAB:Lect UD:Scroll LR:Panneau OK:Add/Ouvr",
-  "hint_np_trk": "TAB:Lect UD:Scroll LR:Panneau OK:Lire DEL:Suppr",
-  "hint_np_pls": "TAB:Lect UD:Scroll LR:Panneau OK:Charg N:Nouv DEL:Suppr",
-  "hint_playlist_sel": "UD:Scroll OK:Charger DEL:Suppr BACK:Menu",
-  "hint_playlist_ed": "LR:Panneau UD:Sel OK:Add/Suppr BACK:Menu",
-  "hint_continue": "ENT:Continuer",
-  "hint_confirm": "LR:Choisir ENT:Confirmer",
-  "confirm": "Confirmer",
-  "delete": "Supprimer"
-}
